spam_check_api/requirements.txt
dj-database-url==2.3.0
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
Faker==30.8.2
gunicorn==23.0.0
python-dateutil==2.9.0.post0
psycopg2==2.9.10
redis==5.2.0
whitenoise==6.8.2